requirements.txt
djangorestframework
django-simple-history